+++ .github/workflows/pull-request.yaml
@@ -33,9 +33,6 @@ jobs:
       - name: Test
         run: |
           make test
-      - name: Test
-        run: |
-          make test
       - name: Upload coverage to Codecov
         uses: codecov/codecov-action@v1
         with:
