+++ Makefile
@@ -10,8 +10,8 @@ build-linux:
 build-all: build-darwin build-linux build-win
 
 build-gui:
+	CGO_ENABLE=0 GOOS=windows go build -ldflags -H=windowsgui -o bin/win/transfer-gui.exe ui/main.go
 	CGO_ENABLE=0 GOOS=windows go build -o bin/win/transfer-gui.exe ui/main.go
-	GOARCH=wasm GOOS=js go build -o bin/web/app.wasm ui/main.go
 
 test:
 	go test ./... -coverprofile coverage.out
