+++ pkg/common_test.go
@@ -1,7 +1,6 @@
-package main
+package pkg
 
 import (
-	"github.com/linuxsuren/transfer/cmd"
 	"testing"
 
 	"github.com/stretchr/testify/assert"
@@ -67,42 +66,6 @@ func TestFillContainerWithNumber(t *testing.T) {
 		t.Run(tt.name, func(t *testing.T) {
 			result := fillContainerWithNumber(tt.num, tt.size)
 			assert.Equal(t, tt.want, result, "failed in case [%d]", i)
-		})
-	}
-}
-
-func TestCheckMissing(t *testing.T) {
-	tests := []struct {
-		name      string
-		message   []byte
-		wantIndex int
-		wantOK    bool
-	}{{
-		name:      "missing",
-		message:   []byte("miss000123"),
-		wantIndex: 123,
-		wantOK:    true,
-	}, {
-		name:      "missing",
-		message:   []byte("miss   123"),
-		wantIndex: 123,
-		wantOK:    true,
-	}, {
-		name:      "missing with invalid number",
-		message:   []byte("miss000aaa"),
-		wantIndex: 0,
-		wantOK:    false,
-	}, {
-		name:      "done",
-		message:   []byte("done0000"),
-		wantIndex: -1,
-		wantOK:    true,
-	}}
-	for i, tt := range tests {
-		t.Run(tt.name, func(t *testing.T) {
-			index, ok := cmd.checkMissing(tt.message)
-			assert.Equal(t, tt.wantIndex, index, "failed in case [%d]", i)
-			assert.Equal(t, tt.wantOK, ok, "failed in case [%d]", i)
 		})
 	}
 }
