+++ pkg/header_test.go
@@ -1,4 +1,4 @@
-package cmd
+package pkg
 
 import (
 	"github.com/stretchr/testify/assert"
