+++ ui/main.go
@@ -1,17 +1,31 @@
 package main
 
 import (
+	"context"
+	_ "embed"
 	"encoding/json"
 	"fmt"
 	"github.com/asticode/go-astikit"
 	"github.com/asticode/go-astilectron"
-	cmd2 "github.com/linuxsuren/transfer/cmd"
+	"github.com/linuxsuren/transfer/pkg"
 	"log"
+	"net/http"
 )
+
+func startHTTPServer() {
+	http.HandleFunc("/", func(w http.ResponseWriter, r *http.Request) {
+		_, _ = w.Write([]byte(indexHTML))
+	})
+	_ = http.ListenAndServe(":9999", nil)
+}
 
 func main() {
 	// Set logger
 	l := log.New(log.Writer(), log.Prefix(), log.Flags())
+
+	go func() {
+		startHTTPServer()
+	}()
 
 	// Create astilectron
 	a, err := astilectron.New(l, astilectron.Options{
@@ -33,14 +47,14 @@ func main() {
 
 	// New window
 	var w *astilectron.Window
-	if w, err = a.NewWindow("index.html", &astilectron.WindowOptions{
+	if w, err = a.NewWindow("http://localhost:9999", &astilectron.WindowOptions{
 		Center: astikit.BoolPtr(true),
 		Height: astikit.IntPtr(700),
 		Width:  astikit.IntPtr(700),
 	}); err != nil {
 		l.Fatal(fmt.Errorf("main: new window failed: %w", err))
 	}
-
+	//w.OpenDevTools()
 	w.OnMessage(func(m *astilectron.EventMessage) (v interface{}) {
 		fmt.Println("receive message")
 		var s string
@@ -54,18 +68,70 @@ func main() {
 
 		switch data["cmd"] {
 		case "wait":
-			cmd := cmd2.NewWaitCmd()
-			cmd.Execute()
+			waiter := pkg.NewUDPWaiter(3000)
+			msg := make(chan string, 10)
+
+			go func() {
+				for m := range msg {
+					var n = a.NewNotification(&astilectron.NotificationOptions{
+						Body:             m,
+						HasReply:         astikit.BoolPtr(true),  // Only MacOSX
+						ReplyPlaceholder: "type your reply here", // Only MacOSX
+						Title:            "Msg",
+					})
+					// Create notification
+					n.Create()
+					// Show notification
+					n.Show()
+				}
+			}()
+			return waiter.Start(msg)
 		case "send":
 			file := data["message"]
 			if file != "" {
-				cmd := cmd2.NewSendCmd()
-				cmd.SetArgs([]string{file})
-				cmd.Execute()
+				sender := pkg.NewUDPSender(data["ip"])
+				msg := make(chan string, 10)
+
+				go func() {
+					for m := range msg {
+						var n = a.NewNotification(&astilectron.NotificationOptions{
+							Body:             m,
+							HasReply:         astikit.BoolPtr(true),  // Only MacOSX
+							ReplyPlaceholder: "type your reply here", // Only MacOSX
+							Title:            "Msg",
+						})
+						// Create notification
+						n.Create()
+						// Show notification
+						n.Show()
+					}
+				}()
+
+				err = sender.Send(msg, file)
 			}
 		}
 		return
 	})
+	_ = pkg.Broadcast(context.TODO())
+
+	ctx, cancel := context.WithCancel(context.TODO())
+	w.On(astilectron.EventNameAppClose, func(e astilectron.Event) (deleteListener bool) {
+		cancel()
+		return
+	})
+	waiter := make(chan string, 10)
+	pkg.FindWaiters(ctx, waiter)
+	go func() {
+		for {
+			select {
+			case <-ctx.Done():
+				return
+			case ip := <-waiter:
+				_ = w.SendMessage(ip, func(m *astilectron.EventMessage) {
+				})
+			}
+		}
+	}()
 
 	// Create windows
 	if err = w.Create(); err != nil {
@@ -75,3 +141,6 @@ func main() {
 	// Blocking pattern
 	a.Wait()
 }
+
+//go:embed index.html
+var indexHTML string
